example/functions/img-rembg-fn-py/requirements.txt
# requests
fastapi
aiohttp
uvicorn
gunicorn
oaas-sdk-py==0.3.3
pillow
rembg
